Show creation with relation test failure
